add async Stream<E> getStream() method and switch command enum test to switch/case
This adds a new method, getStream() that returns a Stream<E> that will iterate on values
as they are produced by a closure running on the background agent/isolate.  If the stream
throws an exception in the background agent, the exception is captured and set in the
agent.error's e and stacktrace variables, and the iteration terminates.

Multiple async streams can be iterated on simultaneously, and the streams can produce
any type <E>, not only the state's <T> type.

I think I covered the basics in the new tests, including one test that confirms values
are being produced "live" using a stopwatch to measure the delay between values from
a Stream<>.periodic(100 millisecond) to ensure it's actually generating values live
from the agent back to the parent, not just collecting everything at once.

I'm new to Dart (and flutter) and I'm exploring the different packages for managing
long-lived isolates for my first app.  Thanks for this package, it's nicely done.
I hope this PR is useful.  Let me know if you'd like any changes.